MOJARRA-749 This changebundle describes the effort to move annotation processing out of the API and into the implementation. The primary motiviator for doing this is the fact that calling any annotation-related method on a Class instances calls through to a syncronized method to init the annotation data for the class.  Repeated calls to these methods would kill scalability.  Additionally the annotation processing was scattered between different parts of the API and implementation which, at least to me, doesn't seem very clean.
Digging in furter:
Previously UIComponentBase, UIInput, and UIOutput performed processing
of @ResourceDependency annotations when a component was added as a
child to the tree.  This was convenient as it meant one could just
create a component (using the application factory or not) and
have it processed.  If by chance the renderer type was set on the component
at this time, any @ResourceDependency annotations on the Renderer
would be processed as well.  This meant that the rendererType must
be set before adding the component as a child.
@ListenerFor annotations would be processed by the implementation.
Any @ListenerFor annotations on the component would be processed
when the application component factory methods would be called, however,
if the component was created via new and then added, the ResourceDependency
annotations would be processed while the @ListenerFor annotations would not
be.  @ListenerFor annotations on the Renderer would be called
by the PDL after the renderer type had been set.
Changes to address the issues:
At a high level, annotation processing will *only* occur when the application
component (beit UIComponent, Converter, or Validator) are used.  No processing
will occur in the API or in the PDL.
New methods added to javax.faces.application.Application:
public UIComponent createComponent(ValueExpression componentExpression,
                                       FacesContext context,
                                       String componentType,
                                       String rendererType);
*and*
public UIComponent createComponent(FacesContext context,
                                   String componentType,
                                   String rendererType);
These methods are currently leveraged by the PDL but would be used when creating
a tree dynamically and cause the annotations on both the UIComponent and
Renderer to be processed.
The older createComponent() methods when invoked will *only* process
component-level annotations.  The Renderer will not be queried in this
case as the default renderer (if the component has one) may not be the one
used after the component is created so we don't want to potentially process
annotations for a renderer that won't be used.
The new classes listed below represent the metadata gathering and caching
of annotations.  This allows us to greatly reduce the number of annotation
method calls on class instances to the first time the cache structure is
created.
SECTION: Modified Files